Refactor url map and parser
