+++ canonicalwebteam/discourse/parsers/__init__.py
@@ -1,0 +1,4 @@
+from canonicalwebteam.discourse.parsers.docsParser import DocParser  # noqa
+from canonicalwebteam.discourse.parsers.engageParser import (  # noqa
+    EngageParser,  # noqa
+)
